[optimize]优化创建时间和更新时间自动填充mybatis字段

+++ tool-server/code-generator/src/main/java/com/github/cloud/enums/DataTypeEnum.java
@@ -17,7 +17,7 @@ public enum DataTypeEnum {
     DATETIME("datetime", "Date"),
     CHAR("char", "String"),
     BIT("bit", "Boolean"),
-    TINYINT("tinyint", "Byte"),
+    TINYINT("tinyint", "Integer"),
     ;
 
     /**
